+++ docs/changelog.rst
@@ -1,8 +1,14 @@
 =========
 CHANGELOG
 =========
+
 2.93.0+dev (XXXX-XX-XX)
 -----------------------
+
+**New feature**
+
+- New ``LEIParser`` to import touristic content and event from LEI touristic data system
+- New ``XMLParser`` to import content from XML
 
 **Bug fixes**